Rename: segmetation -> segutil.
